Update patch set 1
Patch Set 1:

I've seen two kernels that had to resolve null termination of a table before a kernel built with gcc 4.9 would boot. You might check if this is the case for you. Grab these:

https://review.lineageos.org/157874
https://review.lineageos.org/157873

And follow the direction in the comments for 157874, but run the tool on the entire drivers folder. Then check if any of the hits matches a driver you're using.

Patch-set: 1